auth-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
tortoise-orm==0.19.3
asyncpg==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.0
python-dotenv==1.0.0
